requirements.txt
# Core Deep Learning Framework
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1

# Quantum Computing Framework
pennylane==0.39.0
pennylane-lightning==0.39.0

# Scientific Computing & Data Processing
numpy==1.26.4
scipy==1.11.4
pandas==2.1.4

# Signal Processing
mne==1.6.1
pywavelets==1.5.0

# Machine Learning & Metrics
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Configuration & Utilities
pyyaml==6.0.1
tqdm==4.66.1

# Logging
tensorboard==2.16.2

# Data Handling
h5py==3.10.0
